// src/generated/prisma/models/Appointment.ts
export type AppointmentWhereInput = {
  AND?: Prisma.AppointmentWhereInput | Prisma.AppointmentWhereInput[]
  OR?: Prisma.AppointmentWhereInput[]
  NOT?: Prisma.AppointmentWhereInput | Prisma.AppointmentWhereInput[]
  id?: Prisma.UuidFilter<"Appointment"> | string
  doctorId?: Prisma.UuidFilter<"Appointment"> | string
  patientId?: Prisma.UuidFilter<"Appointment"> | string
  serviceName?: Prisma.StringFilter<"Appointment"> | string
  startsAt?: Prisma.DateTimeFilter<"Appointment"> | Date | string
  createdAt?: Prisma.DateTimeFilter<"Appointment"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Appointment"> | Date | string
  doctor?: Prisma.XOR<Prisma.DoctorScalarRelationFilter, Prisma.DoctorWhereInput>
  patient?: Prisma.XOR<Prisma.PatientScalarRelationFilter, Prisma.PatientWhereInput>
}